Updated algorithms for SEARCH and UPDATE
Massively improves SEARCH efficiency at the cost of revealing 'c'
(count of documents that term is found in) in transit and to server.
Server can find out 'c' anyway, but it is visible on the wire.
Shouldn't be a huge issue, especially because each update will alter
that key (representing the term) in the index.

Made some other, simpler changes to the algorithms that improve
performance, but not on the order as that mentioned above.

Additionally, updated time.sh to include command-line args.

+++ client/time.sh
@@ -7,26 +7,38 @@ SEARCH='python sse_client.py -s'
 FILES=('1.' '12.' '15.' '18.' '20.' '23.' '26.' '29.' '31.' '34.') # '37.' '4.' '42.' '45.' '62.' '65.' '68.' '70.' '73.' '78.') # '83.' '86.');
 
 # 10 'known' terms
-TERMS=('such') # 'tuesday' 'pm' 'message' 'if' 'but' 'very' 'problem' 'see' 'the')
+TERMS=( 'tuesday' 'such' 'pm' 'message' 'if' 'but' 'very' 'problem' 'see' 'the')
 
 # 5 'unknown' terms
 UNKNOWN=('beluga' 'whippersnapper' 'cufflink' 'grigsby' 'described')
 
-for i in "${TERMS[@]}"; do
-    echo "$SEARCH $i"
-    $SEARCH $i
-done
-exit 0
-rm index*
-rm ../server/index
-for i in "${FILES[@]}"; do
-    echo "$UPDATE inbox/$i"
-    $UPDATE inbox/$i
-done
-exit 0
+if [ "$1" == "" ]; then
+    echo -e "Must supply a command-line option!\n"
+    echo -e "\t'update' (clears and updates indices with new messages)"
+    echo -e "\t'search' (searches documents for a number of terms)"
+    echo -e "\t'searchfull' (includes terms that will not be found)"
+    echo -e "\t'all' (update and search)"
+    exit 1
+fi
 
+if [ "$1" == "update" ];then
+    rm index*
+    rm ../server/index
+    for i in "${FILES[@]}"; do
+        echo "$UPDATE inbox/$i"
+        $UPDATE inbox/$i
+    done
+fi
+if [ "$1" == "search" ] || [ "$1" == "all" ]; then
 
-for i in "${UNKNOWN[@]}"; do
-    echo "$SEARCH $i";
-    $SEARCH $i
-done
+    for i in "${TERMS[@]}"; do
+        echo "$SEARCH $i"
+        $SEARCH $i
+    done
+fi
+if [ "$1" == "searchfull" ]; then
+    for i in "${UNKNOWN[@]}"; do
+        echo "$SEARCH $i";
+        $SEARCH $i
+    done
+fi
